requirements.txt
faster_whisper
sounddevice
numpy
pyaudio
soxr
webrtcvad-wheels
samplerate
llama-cpp-python